server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "pyyaml~=6.0.2",
  "plotly~=5.24.1",
  "pandas~=2.2.3",
  "fastapi~=0.115.0",
  "ipython~=8.28.0",
  "jinja2~=3.1.2",
  "piccolo-api~=1.5.2",
  "piccolo[all]~=1.22.0",
  "python-dotenv~=1.0.0",
  "uvicorn~=0.32.0",
  "yagmail~=0.15.293",
]
